Move session callback to shared next-auth config
